# webapp/frontend/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "DeploymentCreate": {
        "description": "POST /deployments body.\n\n``name`` defaults to ``\"<ticker>-<strategy>-<train_end>\"`` and\n``warmup_bars`` defaults to the strategy's\n``required_warmup_bars + convergence_margin_bars`` (subject to an\nabsolute floor) when omitted.",
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_kind": {
            "enum": [
              "run",
              "hpo"
            ],
            "title": "Source Kind",
            "type": "string"
          },
          "warmup_bars": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warmup Bars"
          }
        },
        "required": [
          "source_kind",
          "source_id"
        ],
        "title": "DeploymentCreate",
        "type": "object"
      },
      "DeploymentDetail": {
        "description": "Summary + the tail of the signal log.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "interval": {
            "$ref": "#/components/schemas/Interval"
          },
          "latest_signal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SignalRowOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_username": {
            "title": "Owner Username",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_kind": {
            "enum": [
              "run",
              "hpo"
            ],
            "title": "Source Kind",
            "type": "string"
          },
          "strategy_name": {
            "title": "Strategy Name",
            "type": "string"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "train_end": {
            "format": "date-time",
            "title": "Train End",
            "type": "string"
          },
          "warmup_bars": {
            "title": "Warmup Bars",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "source_kind",
          "source_id",
          "ticker",
          "strategy_name",
          "interval",
          "train_end",
          "warmup_bars",
          "created_at",
          "owner_username"
        ],
        "title": "DeploymentDetail",
        "type": "object"
      },
      "DeploymentRename": {
        "description": "PATCH /deployments/{id} body.",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "DeploymentRename",
        "type": "object"
      },
      "DeploymentSummary": {
        "description": "Row in the deployments list view.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "interval": {
            "$ref": "#/components/schemas/Interval"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_username": {
            "title": "Owner Username",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_kind": {
            "enum": [
              "run",
              "hpo"
            ],
            "title": "Source Kind",
            "type": "string"
          },
          "strategy_name": {
            "title": "Strategy Name",
            "type": "string"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "train_end": {
            "format": "date-time",
            "title": "Train End",
            "type": "string"
          },
          "warmup_bars": {
            "title": "Warmup Bars",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "source_kind",
          "source_id",
          "ticker",
          "strategy_name",
          "interval",
          "train_end",
          "warmup_bars",
          "created_at",
          "owner_username"
        ],
        "title": "DeploymentSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HoldoutEvalSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "holdout_start": {
            "format": "date-time",
            "title": "Holdout Start",
            "type": "string"
          },
          "launched_by_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Launched By Username"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sharpe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharpe Ratio"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_kind": {
            "enum": [
              "run",
              "hpo"
            ],
            "title": "Source Kind",
            "type": "string"
          },
          "store": {
            "title": "Store",
            "type": "string"
          }
        },
        "required": [
          "name",
          "store",
          "created_at",
          "source_kind",
          "source_id",
          "holdout_start",
          "sharpe_ratio"
        ],
        "title": "HoldoutEvalSummary",
        "type": "object"
      },
...
      "Interval": {
        "description": "Bar timeframe - used for annualization factor computation.",
        "enum": [
          "second",
          "minute",
          "five_minute",
          "fifteen_minute",
          "hour",
          "daily",
          "weekly"
        ],
        "title": "Interval",
        "type": "string"
      },
...
      "PredictIfStaleResponse": {
        "description": "Response for POST /deployments/{id}/predict-if-stale.\n\n``stale=False`` means the cached signal already covered the latest\navailable bar; ``stale=True`` means a fresh predict ran.",
        "properties": {
          "signal": {
            "$ref": "#/components/schemas/SignalRowOut"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          }
        },
        "required": [
          "stale",
          "signal"
        ],
        "title": "PredictIfStaleResponse",
        "type": "object"
      },
...
      "SignalEvaluationOut": {
        "description": "Response for GET /deployments/{id}/signal-evaluation.\n\nPer-signal scores plus headline stats over the scored subset.\n``hit_rate`` covers directional (non-FLAT) scored signals only;\n``cumulative_return`` compounds every scored ``listened_return``\n(gross), and ``net_cumulative_return`` does so after costs. All summary\nstats are null when nothing has been scored yet. ``cost_scenario`` is\nthe tier whose slippage + commission produced the net figures.",
        "properties": {
          "cost_scenario": {
            "$ref": "#/components/schemas/SlippageScenario"
          },
          "cumulative_return": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cumulative Return"
          },
          "hit_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hit Rate"
          },
          "mean_return": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean Return"
          },
          "n_hits": {
            "title": "N Hits",
            "type": "integer"
          },
          "n_scored": {
            "title": "N Scored",
            "type": "integer"
          },
          "n_signals": {
            "title": "N Signals",
            "type": "integer"
          },
          "net_cumulative_return": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Cumulative Return"
          },
          "net_mean_return": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Mean Return"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ScoredSignalOut"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "rows",
          "n_signals",
          "n_scored",
          "n_hits",
          "hit_rate",
          "cumulative_return",
          "mean_return",
          "net_cumulative_return",
          "net_mean_return",
          "cost_scenario"
        ],
        "title": "SignalEvaluationOut",
        "type": "object"
      },
      "SignalRowOut": {
        "description": "One row from the deployment's signal log.\n\n``bar_ts`` is the last completed bar the signal was computed from;\n``signal_date`` is the trading day the signal is *for* (the next\nsession after ``bar_ts``).",
        "properties": {
          "bar_ts": {
            "format": "date-time",
            "title": "Bar Ts",
            "type": "string"
          },
          "signal": {
            "title": "Signal",
            "type": "number"
          },
          "signal_date": {
            "format": "date-time",
            "title": "Signal Date",
            "type": "string"
          },
          "source_run_id": {
            "title": "Source Run Id",
            "type": "string"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          },
          "warmup_bars_used": {
            "title": "Warmup Bars Used",
            "type": "integer"
          },
          "warmup_fingerprint": {
            "title": "Warmup Fingerprint",
            "type": "string"
          }
        },
        "required": [
          "submitted_at",
          "bar_ts",
          "signal_date",
          "signal",
          "warmup_fingerprint",
          "source_run_id",
          "warmup_bars_used"
        ],
        "title": "SignalRowOut",
        "type": "object"
      },
...
      "SlippageScenario": {
        "description": "Named cost tiers - keys into ``COST_SCENARIOS`` / ``SLIPPAGE_SCENARIOS``.",
        "enum": [
          "zero",
          "low",
          "normal",
          "high"
        ],
        "title": "SlippageScenario",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Quant Webapp",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/deployments": {
      "get": {
        "operationId": "get_deployments_api_deployments_get",
        "parameters": [
          {
            "in": "query",
            "name": "all",
            "required": false,
            "schema": {
              "default": false,
              "title": "All",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeploymentSummary"
                  },
                  "title": "Response Get Deployments Api Deployments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Deployments",
        "tags": [
          "deployments"
        ]
      },
      "post": {
        "operationId": "post_deployment_api_deployments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Deployment",
        "tags": [
          "deployments"
        ]
      }
    },
    "/api/deployments/{deployment_id}": {
      "delete": {
        "operationId": "delete_deployment_route_api_deployments__deployment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "title": "Deployment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Deployment Route",
        "tags": [
          "deployments"
        ]
      },
      "get": {
        "operationId": "get_deployment_detail_api_deployments__deployment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "title": "Deployment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Deployment Detail",
        "tags": [
          "deployments"
        ]
      },
      "patch": {
        "operationId": "patch_deployment_api_deployments__deployment_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "title": "Deployment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentRename"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Deployment",
        "tags": [
          "deployments"
        ]
      }
    },
    "/api/deployments/{deployment_id}/predict-if-stale": {
      "post": {
        "description": "Recall today's signal or compute it now.",
        "operationId": "post_predict_if_stale_api_deployments__deployment_id__predict_if_stale_post",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "title": "Deployment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictIfStaleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Predict If Stale",
        "tags": [
          "deployments"
        ]
      }
    },
    "/api/deployments/{deployment_id}/signal-evaluation": {
      "get": {
        "description": "Score the deployment's emitted signals open->open; ``cost`` sets the net friction tier.",
        "operationId": "get_signal_evaluation_api_deployments__deployment_id__signal_evaluation_get",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "title": "Deployment Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cost",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SlippageScenario",
              "default": "normal"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalEvaluationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Signal Evaluation",
        "tags": [
          "deployments"
        ]
      }
    },
    "/api/deployments/{deployment_id}/signals": {
      "get": {
        "operationId": "get_signal_log_api_deployments__deployment_id__signals_get",
        "parameters": [
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "title": "Deployment Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 10000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SignalRowOut"
                  },
                  "title": "Response Get Signal Log Api Deployments  Deployment Id  Signals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Signal Log",
        "tags": [
          "deployments"
        ]
      }
    },
...
    "/api/holdout-evals": {
      "get": {
        "operationId": "get_holdout_evals_api_holdout_evals_get",
        "parameters": [
          {
            "in": "query",
            "name": "all",
            "required": false,
            "schema": {
              "default": false,
              "title": "All",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HoldoutEvalSummary"
                  },
                  "title": "Response Get Holdout Evals Api Holdout Evals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Holdout Evals",
        "tags": [
          "holdout"
        ]
      }
    },
    "/api/holdout-evals/{name}": {
      "get": {
        "operationId": "get_holdout_eval_detail_api_holdout_evals__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldoutEvalDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Holdout Eval Detail",
        "tags": [
          "holdout"
        ]
      }
    },
    "/api/holdout-evals/{name}/plots/{plot_name}": {
      "get": {
        "operationId": "get_holdout_eval_plot_api_holdout_evals__name__plots__plot_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "plot_name",
            "required": true,
            "schema": {
              "title": "Plot Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Holdout Eval Plot",
        "tags": [
          "holdout"
        ]
      }
    },
...
  }
}